Only add comment to feed if it has title/text
Once in awhile an API response for new comments won't yet contain the
story title the comment is attached to. This leads to RSS feed items
saying the comment appeared in "None."

With this commit, don't show new comment feed items unless both title
and comment text exist.

It usually only lasts a few seconds, so shouldn't be a huge deal to wait.