Dereference numeric character entities
Some hit['story_text'] elements would return a string like '&#x2F;'
which would escaped and included in the XML as '&amp;#x2F;' -- all when
'/' is fine.

This commit replaces all those instances with the proper unicode
character so things look better in the resulting XML.

Semi-related, it also replaces &quot; and &apos; with '"' and '"',
respectively.