Enable searching for multiple terms in a query
If a search query (`q=`) is formatted like "term1 OR term2", the
search results will work as expected.

The way it works is the `query` parameter, in effect, does a "term1 AND term2"
search while the `optionalWords` parameter makes "term1"
optional while looking for "term2" hits, and "term2" optional when
looking for "term1" hits.

Algolia doesn't offer this directly via the `query` parameter (from
what I could find), so we have to get creative.

Refs: #12
Refs: <https://www.algolia.com/doc/rest-api/search/#query>
Refs: <https://www.algolia.com/doc/rest-api/search/#optionalwords>