Make donation link appearance less random
Previously, this could have led to an issue (in theory) where an item
appears updated because the donation link showed up during fetch #1
but not during fetch #2. Or vice versa.

With this commit, pass the guid through a SHA1 hash and if the
hexdigest starts with zero through three (i.e., 4/16 = 25% of the
time, same as before) display the donation link. But now that decision
won't change between fetches.