api: add a 10 second timeout to Algolia API calls
Very rarely Algolia will hang when returning results which causes
hnrss to stop serving feeds as all the threads are tied up.

With this commit, timeout after 10 seconds and if any issue during the
request (timeout or otherwise), just return an empty dict which
generates an empty feed.

An empty feed isn't great, but it's better than a nasty Varnish error.